fix: id String에서 int로 변경

+++ lib/screens/mypage/screen_myComment.dart
@@ -78,7 +78,7 @@ class MyComment extends StatelessWidget {
                     userDetail.photoUrl,
                     userDetail.nickname,
                     boardId as int,
-                    board.userId,
+                    board.userId as int,
                     board.body,
                     board.likeCnt,
                     board.commentCnt,
